Change triggers
'looking for|candidate' is triggering constantly on phrases like 'im just saying how we’ve found good candidates' and 'I have to update my job posting - we’re looking for more of a sysadmin type at this point than a developer type'. 'is hiring' seems more prevalent.

This may under-trigger the bot, but I think that's better that over-triggering it.